+++ src/mako/database/query/compilers/MariaDB.php
@@ -7,12 +7,14 @@
 
 namespace mako\database\query\compilers;
 
+use mako\database\query\compilers\exceptions\CompilerException;
 use mako\database\query\Subquery;
 use mako\database\query\VectorDistance;
 use Override;
 
 use function is_array;
 use function json_encode;
+use function sprintf;
 
 /**
  * Compiles MariaDB queries.
@@ -40,6 +42,11 @@ class MariaDB extends MySQL
 		$function = match ($vectorDistance['vectorDistance']) {
 			VectorDistance::Cosine => 'VEC_DISTANCE_COSINE',
 			VectorDistance::Euclidean => 'VEC_DISTANCE_EUCLIDEAN',
+			default => throw new CompilerException(sprintf(
+				'The [ %s ] query compiler does not support [ %s ] vector distances.',
+				static::class,
+				$vectorDistance['vectorDistance']
+			)),
 		};
 
 		return "{$function}({$this->column($vectorDistance['column'], false)}, {$vector})";