+++ src/mako/database/query/compilers/exceptions/CompilerException.php
@@ -1,0 +1,18 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\database\query\compilers\exceptions;
+
+use mako\database\exceptions\DatabaseException;
+
+/**
+ * Compiler exception.
+ */
+class CompilerException extends DatabaseException
+{
+
+}